requirements.txt
pytest==8.4.2
selenium==4.17.1
requests==2.32.0
allure-pytest==2.15.3
